1 - Reorganização das pastas da UI. 2 - Criação da Work Activity 3 - Transformação parcial do Fragment Menu em Activity
