+++ maxApp/app/build.gradle
@@ -53,6 +53,6 @@ dependencies {
     // ViewModel and LiveData
     implementation "androidx.lifecycle:lifecycle-extensions:$lifecycle_version"
 
-    def anko_version='0.10.8'
+    def anko_version = '0.10.8'
     implementation "org.jetbrains.anko:anko:$anko_version"
 }
