+++ maxApp/app/src/main/java/br/com/mauriciobenigno/maxapp/ui/Fragments/Dados/DadosDoClienteViewModel.kt
@@ -1,4 +1,4 @@
-package br.com.mauriciobenigno.maxapp.ui
+package br.com.mauriciobenigno.maxapp.ui.Fragments.Dados
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
