+++ maxApp/app/src/main/java/br/com/mauriciobenigno/maxapp/ui/Fragments/HistoricoPedidos/HistoricoDePedidosViewModel.kt
@@ -1,4 +1,4 @@
-package br.com.mauriciobenigno.maxapp.ui
+package br.com.mauriciobenigno.maxapp.ui.Fragments.HistoricoPedidos
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
