+++ maxApp/app/src/main/java/br/com/mauriciobenigno/maxapp/ui/Fragments/Menu/MenuPrincipalFragment.kt
@@ -1,5 +1,6 @@
-package br.com.mauriciobenigno.maxapp.ui
+package br.com.mauriciobenigno.maxapp.ui.Fragments.Menu
 
+import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -7,6 +8,7 @@ import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import br.com.mauriciobenigno.maxapp.R
+import br.com.mauriciobenigno.maxapp.ui.Activitys.WorkActivity
 import kotlinx.android.synthetic.main.fragment_menu_principal.*
 
 class MenuPrincipalFragment : Fragment(){
@@ -24,6 +26,9 @@ class MenuPrincipalFragment : Fragment(){
 
         btnClientes.setOnClickListener {
             Toast.makeText(it.context,"TESTE", Toast.LENGTH_SHORT).show()
+            //findNavController().navigate(R.id.action_menuPrincipalFragment_to_dadosDoClienteFragment)
+            val intent = Intent(this.context, WorkActivity::class.java)
+            activity!!.startActivity(intent)
         }
     }
 }
